Split out dev actions into controllers
Regular request actions and event-streaming actions can't be intermixed
in the same controller. It also seems tidier to have this functionality
in separate controllers, so I've just broken them all out into their own
controllers with common behaviour in a concern.
