+++ app/controllers/dev/reset_organisation_controller.rb
@@ -1,27 +1,10 @@
 # frozen_string_literal: true
 
-class DevController < ApplicationController
+class Dev::ResetOrganisationController < ApplicationController
   include ActionController::Live
+  include DevConcern
 
-  skip_before_action :authenticate_user!
-  skip_before_action :store_user_location!
-  skip_after_action :verify_policy_scoped
-
-  before_action :ensure_dev_env_or_dev_tools_enabled
-
-  def reset
-    session.delete :user_return_to
-    Rake::Task.clear
-    Rails.application.load_tasks
-
-    Organisation.with_advisory_lock("reset") do
-      Rake::Task["db:seed:replant"].invoke
-    end
-
-    redirect_to root_path
-  end
-
-  def reset_organisation
+  def call
     response.headers["Content-Type"] = "text/event-stream"
     response.headers["Cache-Control"] = "no-cache"
 
@@ -79,42 +62,13 @@ class DevController < ApplicationController
     end
 
     response.stream.write "Done"
+  rescue StandardError => e
+    response.stream.write "Error: #{e.message}\n"
   ensure
     response.stream.close
   end
 
-  def random_consent_form
-    Faker::Config.locale = "en-GB"
-
-    session = Session.includes(programmes: :vaccines).find(params[:session_id])
-
-    attributes =
-      if ActiveModel::Type::Boolean.new.cast(params[:parent_phone])
-        {}
-      else
-        { parent_phone: nil }
-      end
-
-    consent_form =
-      FactoryBot.create(:consent_form, :draft, session:, **attributes)
-
-    consent_form.seed_health_questions
-    consent_form.each_health_answer do |health_answer|
-      health_answer.response = "no"
-    end
-    consent_form.save!
-
-    request.session[:consent_form_id] = consent_form.id
-    redirect_to confirm_parent_interface_consent_form_path(consent_form)
-  end
-
   private
-
-  def ensure_dev_env_or_dev_tools_enabled
-    unless Rails.env.local? || Flipper.enabled?(:dev_tools)
-      raise "Not in development environment"
-    end
-  end
 
   def log_destroy(query)
     where_clause = query.where_clause
